Started that long journey...
Now programm manipulate with Locations.itdb. Suddenly, it crashes
whole music on iPod. Probably that's because of
Locations.itdb.cbk (idk what is that tbh, further work is needed to
figure it out), or maybe because of we didn't change Library.itdb
somehow, so iPod see that there's some new file on it, but there's no
this file in Library.itdb and it gets dumb :(